+++ lib/src/utils/Is_equal_validation.dart
@@ -1,0 +1,29 @@
+import 'package:queen_validators/queen_validators.dart';
+
+/// Match Validator for text fields
+/// checks if the input matches other string
+class IsEqual extends TextValidationRule {
+  final String Function() other;
+  bool caseSensitive;
+  String errorMsg;
+
+  IsEqual(
+      {required this.other,
+      this.errorMsg = 'There is no match',
+      this.caseSensitive = true})
+      : super(errorMsg);
+
+  ///  return a bool Either valid or NOT
+  @override
+  bool isValid(String input) {
+    return caseSensitive
+        ? _match(input, other())
+        : _match(input.toLowerCase(), other().toLowerCase());
+  }
+}
+
+bool _match(
+  String input,
+  String other,
+) =>
+    input == other;