+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.1.0
+- Add support for Dart 3.0.0 and Flutter 3.10
+- Add new `XColorScheme` in `playx_theme` which is a new way to configure custom colors for each themes.
+- update packages
+- Update `Playx.runPlayx` function to support Sentry.
+- Add new packages like `Lottie`, `Sentry`, `Async` and `Fimber`
+- Add `IsEqual` which validates text form fields.
+- Update example project.
+
 ## 0.0.9
 - hide `RContextMediaQuery` from `flutter_readable` package to avoid conflicts with `Getx` package.
 - update exports.